% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{allele_numbers}
\alias{allele_numbers}
\title{A slowpop function}
\usage{
allele_numbers(population, n_alleles = NA)
}
\arguments{
\item{population}{is the population}
}
\description{
Get the how many copies there are of each haplotype
}
\examples{
my_population <- create_population(100)
allele_numbers(my_population)
}
\keyword{alleles}
\keyword{genotypes}
\keyword{haplotypes}

