% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{assign_relative_fitness}
\alias{assign_relative_fitness}
\title{A slowpop function}
\usage{
assign_relative_fitness(population, relative_fitness)
}
\arguments{
\item{population}{is the population}

\item{relative_fitness}{a named vector with the relative fitness of each haplotype}
}
\description{
Calculates the relative probability of each individual to make it into the next generation
}
\examples{
my_population <- create_population(100)
my_relative_fitness <- get_named_vector(2)
my_relative_fitness <- c(1,1.1,1.1)
my_probabilities <- assign_relative_fitness(my_population,my_relative_fitness)
}
\keyword{population}
\keyword{test}

