% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{create_population}
\alias{create_population}
\title{A slowpop function}
\usage{
create_population(n, diploid = TRUE, haplotypes = 2)
}
\arguments{
\item{n}{number of individuals in the population}

\item{diploid}{is the organism diploid or not (TRUE/FALSE)}

\item{haplotypes}{the number of haplotypes in the population}
}
\description{
This function creates a population
}
\examples{
my_population <- create_population(100)
}
\keyword{population}

