% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{genotype_numbers}
\alias{genotype_numbers}
\title{A slowpop function}
\usage{
genotype_numbers(population)
}
\arguments{
\item{population}{is the population}
}
\description{
Get the number of individuals for each genotype
}
\examples{
my_population <- create_population(100)
genotype_numbers(my_population)
}
\keyword{genotypes}
\keyword{haplotypes}

