% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_individuals}
\alias{get_individuals}
\title{A slowpop function}
\usage{
get_individuals(relative_fitness, n = 1)
}
\arguments{
\item{relative_fitness}{is a vector with the relative fitness of each individual}

\item{population}{is the population}
}
\description{
Select individuals based on relative fitness
}
\examples{
my_population <- create_population(100)
my_relative_fitness <- get_named_vector(2)
my_relative_fitness <- c(1,1.1,1.1)
my_probabilities <- assign_relative_fitness(my_population,my_relative_fitness)
}
\keyword{population}
\keyword{test}

