% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_named_vector}
\alias{get_named_vector}
\title{A slowpop function}
\usage{
get_named_vector(n = NA)
}
\arguments{
\item{n}{is the number of haplotypes}
}
\description{
Get a vector for each combination of n haplotypes
}
\examples{
my_population <- create_population(100)
get_named_vector(2)
}
\keyword{alleles}
\keyword{genotypes}
\keyword{haplotypes}

