% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{introduce_mutation}
\alias{introduce_mutation}
\title{A slowpop function}
\usage{
introduce_mutation(population)
}
\arguments{
\item{population}{is the population}
}
\description{
Introduce a new haplotype in the population by changing one haplotype in one individual. The haplotype get the maximum number of any existing haplotype plus one.
}
\examples{
my_population <- create_population(100)
introduce_mutation(my_population)
}
\keyword{homozygotes}

