% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{max_haplotype_no}
\alias{max_haplotype_no}
\title{A slowpop function}
\usage{
max_haplotype_no(population)
}
\arguments{
\item{population}{is the population}
}
\description{
Get the highest number assigned to a haplotype in the population
}
\examples{
my_population <- create_population(100)
max_haplotype_no(my_population)
}
\keyword{homozygotes}

