% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{n_haplotypes}
\alias{n_haplotypes}
\title{A slowpop function}
\usage{
n_haplotypes(population)
}
\arguments{
\item{population}{is the population}
}
\description{
Get the number of haplotypes in the population
}
\examples{
my_population <- create_population(100)
n_haplotypes(my_population)
}
\keyword{homozygotes}

