% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{n_homozygots}
\alias{n_homozygots}
\title{A slowpop function}
\usage{
n_homozygots(population)
}
\arguments{
\item{population}{is the population}
}
\description{
Get the number of homozygotes in the population
}
\examples{
my_population <- create_population(100)
n_homozygots(my_population)
}
\keyword{homozygotes}

