% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{new_generation}
\alias{new_generation}
\title{A slowpop function}
\usage{
new_generation(population, relative_fitness = NA)
}
\arguments{
\item{population}{is the population}

\item{relative_fitness}{a named vector with the relative fitness of each haplotype}
}
\description{
Generates a new generation from a given population
}
\examples{
my_population <- create_population(100)
my_relative_fitness <- get_named_vector(2)
my_relative_fitness <- c(1,1.1,1.1)
my_new_generation <- new_generation(my_population,my_relative_fitness)
}
\keyword{population}

