% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{test_population}
\alias{test_population}
\title{A slowpop function}
\usage{
test_population(population)
}
\arguments{
\item{population}{is the population to test}
}
\description{
This function test if the population exist
}
\examples{
my_population <- create_population(100)
test_population(my_population)
}
\keyword{population}
\keyword{test}

